// src/shared/api/generated/models/examinationManagePluginV2ViewVInsertUserOrder.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */

export interface ExaminationManagePluginV2ViewVInsertUserOrder {
  examTypeId?: string;
  /** 输入的好友邀请码 */
  iptInviteCode?: string;
  province?: string;
  provinceCode?: string;
}